+++ src/main/java/es/geeko/repository/TransaccionRepository.java
@@ -3,5 +3,5 @@ package es.geeko.repository;
 import es.geeko.model.Transaccion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransaccionRepository extends JpaRepository<Transaccion,Long> {
+public interface TransaccionRepository extends JpaRepository<Transaccion,Integer> {
 }
